corrections
class AnalyticsCounter modified
create main class Analytic
