+++ analytics/result.out
@@ -1,5 +1,5 @@
 Frequency of amnesia is 1
-Frequency of anxiety is 5
+Frequency of anxiety is 6
 Frequency of arrhythmias is 3
 Frequency of blindness is 1
 Frequency of blurred vision is 5
@@ -15,6 +15,7 @@ Frequency of inflamation is 7
 Frequency of insomnia is 4
 Frequency of liver problem is 1
 Frequency of low blood pressure is 4
+Frequency of mal de dos is 1
 Frequency of nausea is 5
 Frequency of rapid heart rate is 1
 Frequency of rash is 4
