+++ analytics/src/com/hemebiotech/analytics/Analytic.java
@@ -1,0 +1,28 @@
+package com.hemebiotech.analytics;
+
+
+import java.util.List;
+
+
+public class Analytic {
+	/**
+	 * to count occurrences of elements of file  
+	 *
+	 */	
+	
+	public static void main(String args[]) throws Exception {
+		
+		String filepath= "symptoms.txt";
+		
+		ReadSymptomDataFromFile symptomFile = new ReadSymptomDataFromFile(filepath);
+		
+		List<String> names = symptomFile.getSymptoms();
+		
+		AnalyticsCounter nameList = new AnalyticsCounter();
+		
+		nameList.countSymptoms(names);
+			    
+	 }	
+
+}
+
