+++ analytics/src/com/hemebiotech/analytics/ISymptomReader.java
@@ -1,6 +1,5 @@
 package com.hemebiotech.analytics;
 
-import java.io.IOException;
 import java.util.List;
 
 /**
@@ -17,8 +16,7 @@ public interface ISymptomReader {
 	 * 
 	 * @return a raw listing of all Symptoms obtained from a data source, duplicates are possible/probable
 	 */
-	List<String> GetSymptoms ();
-	void countSymptoms(List<String> symptoms) throws IOException;
+	List<String> getSymptoms ();
 	
 }
